requirements.txt
django>=3.2,<4.0
firebase-admin>=5.0.0
pyrebase4>=4.5.0
django-cors-headers>=3.7.0
django-rest-framework>=3.15.2
pwnedpasswords>=2.0.0
google-api-python-client>=2.154.0
google-auth-httplib2>=1.2.1
google-auth-oauthlib>=2.0.0